+++ files.py
@@ -32,8 +32,8 @@ def view_file_properties(path, filename):
 
 def main():
     if len(sys.argv) >= 2:
-        if sys.argv[1] >= "1.9.0":
-            ekernel.splashScreen("ProcyonCLS File Explorer", "Version 1.9.0")
+        if sys.argv[1] >= "1.9.1":
+            ekernel.splashScreen("ProcyonCLS File Explorer", "Version 1.9.1")
             ekernel.printHeader("File Explorer")
             current_path = os.getcwd()
             while True: