+++ notes.py
@@ -5,8 +5,8 @@ import os
 
 def main():
     if len(sys.argv) >= 2:
-        if sys.argv[1] >= "1.9.0":
-            ekernel.splashScreen("ProcyonCLS Notes", "Version 1.9.0")
+        if sys.argv[1] >= "1.9.1":
+            ekernel.splashScreen("ProcyonCLS Notes", "Version 1.9.1")
             ekernel.printHeader("Notes")
             kernel.printInfo("● R - Read")
             kernel.printInfo("● W - Write")