+++ updater.py
@@ -106,7 +106,7 @@ def replaceLocalFiles(extracted_path, target_path):
 def main():
     if len(sys.argv) >= 2:
         if sys.argv[1] != None:
-            ekernel.splashScreen("ProcyonCLS Updater", "Version 1.9.0")
+            ekernel.splashScreen("ProcyonCLS Updater", "Version 1.9.1")
             ekernel.printHeader("ProcyonCLS Updater")
             current_tag = readCurrentTag()
             kernel.printInfo(f"Current version: {current_tag}")
@@ -116,11 +116,11 @@ def main():
             time.sleep(2)
             latest_tag, zip_url = getLatestReleaseTag()
             if latest_tag > current_tag:
-                kernel.printInfo(f"Update available: {current_tag} -> {latest_tag}")
+                kernel.printInfo(f"Update available: {latest_tag}")
                 print()
                 kernel.println(fetchWhatsNew())
                 print()
-                confirm = input("Do you want to update? (y/n) : ").strip()
+                confirm = input(f"Do you want to update? {current_tag} -> {latest_tag} (y/n) : ").strip()
                 if confirm.lower() != "y":
                     kernel.printWarning("Update cancelled by user")
                 elif confirm.lower() == "y":
@@ -128,9 +128,9 @@ def main():
                     kernel.printInfo("Updating ProcyonCLS...")
                     zip_path = os.path.join(current_directory, "latest_release.zip")
                     temp_extract_path = os.path.join(current_directory, "temp")
-                    kernel.println("● Downloading update..")
+                    kernel.println(" ● Downloading update..")
                     downloadRelease(zip_url, zip_path)
-                    kernel.println("● Extracting update..")
+                    kernel.println(" ● Extracting update..")
                     extractRelease(zip_path, temp_extract_path)
 
                     extracted_folder_name = None
@@ -141,11 +141,11 @@ def main():
 
                     if extracted_folder_name:
                         extracted_path = os.path.join(temp_extract_path, extracted_folder_name)
-                        kernel.println("● Updating files...")
+                        kernel.println(" ● Updating files...")
                         replaceLocalFiles(extracted_path, current_directory)
-                        kernel.println("● Writing new tag..")
+                        kernel.println(" ● Writing new tag..")
                         writeCurrentTag(latest_tag)
-                        kernel.println("● Cleaning up..")
+                        kernel.println(" ● Cleaning up..")
                         shutil.rmtree("temp")
                         os.remove("latest_release.zip")
                         kernel.printSuccess("Update completed successfully!")
@@ -159,7 +159,7 @@ def main():
                 kernel.printWarning("You're using version newer than version published, make sure you obtained current version from trusted sources")
             else:
                 kernel.printSuccess("You're up to date!")
-                os.execv(sys.executable, ['python3', 'shell.py', '1.9.0'])
+                os.execv(sys.executable, ['python3', 'shell.py', '1.9.1'])
         else:
             kernel.printError("This version of updater is incompatible with the current version of ProcyonCLS")
     else: